core | taskList performance update & moving functions getting data from deft to separate lib
