+++ core/common/models.py
@@ -1040,112 +1040,3 @@ class ProductionTask(models.Model):
     class Meta:
         #db_table = u'T_PRODUCTION_STEP'
         db_table = u'T_PRODUCTION_TASK'
-
-
-class MCPattern(models.Model):
-    STEPS = ['Evgen',
-             'Simul',
-             'Merge',
-             'Digi',
-             'Reco',
-             'Rec Merge',
-             'Rec TAG',
-             'Atlfast',
-             'Atlf Merge',
-             'Atlf TAG']
-    STATUS = [(x,x) for x in ['IN USE','Obsolete']]
-    id =  models.DecimalField(decimal_places=0, max_digits=12, db_column='mcp_id', primary_key=True)
-    pattern_name =  models.CharField(max_length=150, db_column='pattern_name', unique=True)
-    pattern_dict = models.CharField(max_length=2000, db_column='pattern_dict')
-    pattern_status = models.CharField(max_length=20, db_column='pattern_status', choices=STATUS)
-
-    def save(self, *args, **kwargs):
-        if not self.id:
-            self.id = prefetch_id('deft',u'ATLAS_DEFT.T_PRODUCTION_MCP_ID_SEQ','T_PRODUCTION_MC_PATTERN','MCP_ID')
-        super(MCPattern, self).save(*args, **kwargs)
-
-    class Meta:
-        db_table = u'T_PRODUCTION_MC_PATTERN'
-
-
-class MCPriority(models.Model):
-    STEPS = ['Evgen',
-             'Simul',
-             'Simul(Fast)',
-             'Merge',
-             'Digi',
-             'Reco',
-             'Rec Merge',
-             'Rec TAG',
-             'Atlfast',
-             'Atlf Merge',
-             'Atlf TAG']
-    id = models.DecimalField(decimal_places=0, max_digits=12, db_column='mcprior_id', primary_key=True)
-    priority_key = models.DecimalField(decimal_places=0, max_digits=12, db_column='priority_key', unique=True)
-    priority_dict = models.CharField(max_length=2000, db_column='priority_dict')
-
-    def save(self, *args, **kwargs):
-        if self.priority_key == -1:
-            return
-        if not self.id:
-            self.id = prefetch_id('deft',u'ATLAS_DEFT.T_PRODUCTION_MCPRIOR_ID_SEQ','T_PRODUCTION_MC_PRIORITY','MCPRIOR_ID')
-        super(MCPriority, self).save(*args, **kwargs)
-
-    def priority(self, step, tag):
-        priority_py_dict = json.loads(self.priority_dict)
-        if step == 'Simul' and tag[0] == 'a':
-            step = 'Simul(Fast)'
-        if step in priority_py_dict:
-            return priority_py_dict[step]
-        else:
-            raise LookupError('No step %s in priority dict' % step)
-
-    class Meta:
-        db_table = u'T_PRODUCTION_MC_PRIORITY'
-
-
-def get_priority_object(priority_key):
-    try:
-        mcp = MCPriority.objects.get(priority_key=priority_key)
-    except ObjectDoesNotExist:
-        priority_py_dict = {}
-        for step in MCPriority.STEPS:
-            priority_py_dict.update({step:int(priority_key)})
-        mcp=MCPriority.objects.create(priority_key=-1,priority_dict=json.dumps(priority_py_dict))
-    except Exception as e:
-        raise e
-    return mcp
-
-
-def get_default_nEventsPerJob_dict():
-    defult_dict = {
-        'Evgen':5000,
-        'Simul':100,
-        'Merge':1000,
-        'Digi':500,
-        'Reco':500,
-        'Rec Merge':5000,
-        'Rec TAG':25000,
-        'Atlfast':500,
-        'Atlf Merge':5000,
-        'Atlf TAG':25000
-    }
-    return defult_dict
-
-
-def get_default_project_mode_dict():
-    default_dict = {
-         'Evgen':'spacetoken=ATLASDATADISK',
-         'Simul':'spacetoken=ATLASDATADISK',
-         'Merge':'spacetoken=ATLASMCTAPE',
-         'Digi':'Npileup=5;spacetoken=ATLASDATADISK',
-         'Reco':'Npileup=5;spacetoken=ATLASDATADISK',
-         'Rec Merge':'spacetoken=ATLASDATADISK',
-         'Rec TAG':'spacetoken=ATLASDATADISK',
-         'Atlfast':'Npileup=5;spacetoken=ATLASDATADISK',
-         'Atlf Merge':'spacetoken=ATLASDATADISK',
-         'Atlf TAG':'spacetoken=ATLASDATADISK'
-    }
-    return default_dict
-
-
