--- core/templates/solrdatarefresh.txt
@@ -1,17 +1,0 @@
-
-{% extends "_base_core.html" %}
-
-{% block page_title %} PanDA Monitor Help{% endblock %}
-{% block title %} {{ viewParams.MON_VO }} PanDA monitor{% endblock %}
-{% block subtitle %} Consolidated help information {% endblock %}
-
-{% block body %}
-<table style="margin:10px; width:80%" >
-<tr class="docsection"><th> PanDA Monitor Help </th></tr>
-<tr><td class="doctextlight">
-
-<p>
-This page consolidates information about data refresh in Sorl.
-
-
-{% endblock %}
